hugetlb: add HGM support for hugetlb_fault and hugetlb_no_page
This CL is the first main functional HugeTLB change. Together, these
changes allow the HugeTLB fault path to handle faults on HGM-enabled
VMAs. The two main behaviors that can be done now:
  1. Faults can be passed to handle_userfault.
  2. Faults on pages that have been partially mapped (and userfaultfd is
     not being used) will get mapped at the largest possible size.
     For example, if a 1G page has been partially mapped at 2M, and we
     fault on an unmapped 2M section, hugetlb_no_page will create a 2M
     PMD to map the faulting address.

This commit does not handle hugetlb_wp right now, and it doesn't handle
HugeTLB page migration and swap entries.

huge_pte_alloc_high_granularity be used to do a HugeTLB page table walk
where we may need to split a leaf-level huge PTE into a new page table
level.

Consider the case where we want to install 4K inside an empty 1G page:
1. We walk to the PUD and notice that it is pte_none.
2. We split the PUD by calling `hugetlb_split_to_shift`, creating a
   standard PUD that points to PMDs that are all pte_none.
3. We continue the PT walk to find the PMD. We split it just like we
   split the PUD.
4. We find the PTE and give it back to the caller.

To avoid concurrent splitting operations on the same page table entry,
we require that the mapping rwsem is held for writing while collapsing
and for reading when doing a high-granularity PT walk.

This commit also introduces the function that performs a PT split for
HugeTLB: hugetlb_split_to_shift. It will optimally split the page
table to map a particular address at a particular granularity.
"Optimally" means that we split only as much as is necessary. As with
the example above, if we need to split a 1G PUD to get a 4K PTE out,
then we split the 1G PUD into 2M PMDs, and then we split the 2M PMD that
would hold our 4K PTE into 4K PTEs.

Change-Id: Ic514e3efa101457fc9a3e2d9bbde0f904d96529d