hugetlb: use struct hugetlb_pte for walk_hugetlb_range
Although this change is large, it is somewhat straightforward. Before,
all users of walk_hugetlb_range could get the size of the PTE just be
checking the hmask or the mm_walk struct. With HGM, that information is
held in the hugetlb_pte struct, so we provide that instead of the raw
pte_t*.

Change-Id: I570bec839f2d35ac8a9fcb46795ada045453eeeb