hugetlb: add HGM support for copy_hugetlb_page_range
This allows fork() to work with high-granularity mappings. The page
table structure is copied such that partially mapped regions will remain
partially mapped in the same way for the new process.

A page's reference count is incremented for *each* portion of it that is
mapped in the page table. For example, if you have a PMD-mapped 1G page,
the reference count will be incremented by 512.

Mapcount is handled similar to THPs: if we're mapping the entire
compound page, increment the folio's compound mapcount. If we're mapping
part of the folio, increment the mapcount for each 4K page that is being
mapped. For example, if a 2M piece of a 1G hugepage is being mapped with
a single PMD, each 4K page within the 2M piece will have its mapcount
incremented once.

Change-Id: I2913e8736f10a71dcec1789776aec5e6ef743f7c