hugetlb: add HGM support for copy_hugetlb_page_range
This allows fork() to work with high-granularity mappings. The page
table structure is copied such that partially mapped regions will remain
partially mapped in the same way for the new process.

A page's reference count is incremented for *each* portion of it that is
separately mapped in the page table. For example, if you have a
PMD-mapped 1G page, the reference count and mapcount will be incremented
by 512.

Change-Id: I2913e8736f10a71dcec1789776aec5e6ef743f7c