hugetlb: update follow_hugetlb_page to support HGM
This enables support for GUP, and it is needed for the KVM demand paging
self-test to work.

One important change here is that, before, we never needed to grab the
i_mmap_sem, but now, to prevent someone from collapsing the page tables
out from under us, we grab it for reading when doing high-granularity PT
walks.

Change-Id: I764935513647b48e8789331a7089786a6934b8de