hugetlb: update page_vma_mapped to do high-granularity walks
This updates the HugeTLB logic to look a lot more like the PTE-mapped
THP logic. When a user calls us in a loop, we will update pvmw->address
to walk to each page table entry that could possibly map the hugepage
containing pvmw->pfn.

This makes use of the new pte_order so callers know what size PTE
they're getting.

Change-Id: Iac32d8c993b840235a9a0aeec0c5438a5c978239