hugetlb: extend vma lock for shared vmas
This allows us to add more data into the shared structure, which we will
use to store whether or not HGM is enabled for this VMA or not, as HGM
is only available for shared mappings.

It may be better to include HGM as a VMA flag instead of extending the
VMA lock structure.

Change-Id: Idb89a316a1799eded9f9fedcf655074242f17027