hugetlb: don't set PageUptodate for UFFDIO_CONTINUE
If would be bad if we actually set PageUptodate with UFFDIO_CONTINUE;
PageUptodate indicates that the page has been zeroed, and we don't want
to give a non-zeroed page to the user.

The reason this change is being made now is because UFFDIO_CONTINUEs on
subpages definitely shouldn't set this page flag on the head page.

Change-Id: I8c7ec4f1d77ad260993e037b123098bf16cfa152