selftests/mm: add hugetlb HGM test to migration selftest
This is mostly the same as the shared HugeTLB case, but instead of
mapping the page with a regular page fault, we map it with lots of
UFFDIO_CONTINUE operations. We also verify that the contents haven't
changed after the migration, which would be the case if the
post-migration PTEs pointed to the wrong page.

Change-Id: I3b74a8e304799d9009c6e3161ab71b5e8ed4afb2