hugetlbfs: for unmapping, treat HGM-mapped pages as potentially mapped
hugetlb_vma_maps_page was mostly used as an optimization: if the VMA
isn't mapping a page, then we don't have to attempt to unmap it again.
We are still able to call the unmap routine if we need to.

For high-granularity mapped pages, we can't easily do a full walk to see
if the page is actually mapped or not, so simply return that it might
be.

Change-Id: I25c00e21099a8e460dbdbd3601d28fedc5bc7c5b