hugetlbfs: do a full walk to check if vma maps a page
Because it is safe to do so, we can do a full high-granularity page
table walk to check if the page is mapped.

If it were not safe to do so, we could bail out early in the case of a
high-granularity mapped PTE, indicating that the page could have been
mapped.

Change-Id: I25c00e21099a8e460dbdbd3601d28fedc5bc7c5b