rmap: in try_to_{migrate,unmap}_one, check head page for page flags
The main complication here is that HugeTLB pages have their poison
status stored in the head page as the HWPoison page flag. Because
HugeTLB high-granularity mapping can create PTEs that point to subpages
instead of always the head of a hugepage, we need to check the
compound_head for page flags.

Change-Id: Ib4853fafe62ed1df2eed2f84ac41cc1b7a13756d