hugetlb: rename __vma_shareable_flags_pmd to __vma_has_hugetlb_vma_lock
Previously, if the hugetlb VMA lock was present, that meant that the VMA
was PMD-shareable. Now it is possible that the VMA lock is allocated but
the VMA is not PMD-shareable: if the VMA is a high-granularity VMA.

It is possible for a high-granularity VMA not to have a VMA lock; in
this case, MADV_COLLAPSE will not be able to collapse the mappings.

Change-Id: I34513c0d91b501dd3e4c8dd1ebc8ceecc3d59510